Add register accesses cocotb tests
This commit introduces the required components to verify the
register accesses using cocotb.

A master interface relying on cocotb-bus is defined to simplify
the read and write operations and make them repeatable. It is
extendable to support a toggling of the strobe signal.

The actual test is built with random values and valid addresses
sent to the register interface. The status is pass with this test
only.

More tests can be added in the future to verify the strobe
behavior, the error signal when invalid addresses are sent, and
finally run the test for different DataWidth and ByteAlign values
using pytest.