Redefine simple register interface
To prepare for next implementations of the register interface that
will comply to real protocols, the current one following a simple
logic is renamed to "simple_reg_interface".

Also, to be more realistic, the response data is delayed of one
cycle so that every transaction lasts atleast two cycles.
The testbench is also slightly modified to only test realistic
values for the block to be computed: 512 and 1024 bits.

The register interface is heavily refactored to support read-only
registers for the digest value. The memory map is different and
well defined with this update but the overall functionality and
genericity remains the same.