Randomly constrained tests for register interface
Register interface stimuli is now driven by random and constrained
Request objects, using PyVSC. The tests were all combined into one
that is repeated several times with the random requests.

Coverage is gathered for the requests, until responses are also
sampled to get a better overview of the test runs.